+++ noodles-bam/src/record/codec/encoder/cigar.rs
@@ -40,24 +40,22 @@ impl fmt::Display for EncodeError {
 // § 4.2.2 "N_CIGAR_OP field" (2023-11-16): "For an alignment with more [than 65535] CIGAR
 // operations, BAM [...] sets `CIGAR` to `kSmN` as a placeholder, where `k` equals `l_seq`, `m` is
 // the reference sequence length in the alignment..."
-pub(super) fn overflowing_write_cigar_op_count<C>(
+pub(super) fn overflowing_write_cigar_op_count(
     dst: &mut Vec<u8>,
     base_count: usize,
-    cigar: &C,
-) -> io::Result<Option<sam::alignment::record_buf::Cigar>>
-where
-    C: Cigar,
-{
+    op_count: usize,
+    alignment_span: usize,
+) -> io::Result<Option<sam::alignment::record_buf::Cigar>> {
     const OVERFLOWING_OP_COUNT: u16 = 2;
 
-    if let Ok(op_count) = u16::try_from(cigar.len()) {
-        write_u16_le(dst, op_count);
+    if let Ok(n) = u16::try_from(op_count) {
+        write_u16_le(dst, n);
         Ok(None)
     } else {
         write_u16_le(dst, OVERFLOWING_OP_COUNT);
 
         let k = base_count;
-        let m = cigar.alignment_span()?;
+        let m = alignment_span;
 
         Ok(Some(
             [Op::new(Kind::SoftClip, k), Op::new(Kind::Skip, m)]
@@ -111,7 +109,6 @@ where
 
 #[cfg(test)]
 mod tests {
-    use std::iter;
 
     use noodles_sam::alignment::record_buf::Cigar as CigarBuf;
 
@@ -122,16 +119,14 @@ mod tests {
         let mut buf = Vec::new();
 
         buf.clear();
-        let cigar: CigarBuf = [Op::new(Kind::Match, 4)].into_iter().collect();
-        let overflowing_cigar = overflowing_write_cigar_op_count(&mut buf, 4, &cigar)?;
+        let overflowing_cigar = overflowing_write_cigar_op_count(&mut buf, 4, 1, 4)?;
         assert_eq!(buf, [0x01, 0x00]);
         assert!(overflowing_cigar.is_none());
 
         const MAX_CIGAR_OP_COUNT: usize = (1 << 16) - 1;
         buf.clear();
-        let cigar: CigarBuf =
-            iter::repeat_n(Op::new(Kind::Match, 1), MAX_CIGAR_OP_COUNT + 1).collect();
-        let overflowing_cigar = overflowing_write_cigar_op_count(&mut buf, 4, &cigar)?;
+        let overflowing_cigar =
+            overflowing_write_cigar_op_count(&mut buf, 4, MAX_CIGAR_OP_COUNT + 1, 1 << 16)?;
         assert_eq!(buf, [0x02, 0x00]);
         assert_eq!(
             overflowing_cigar,